Adding changes requested by Gene: On intro to TM, it says “Integrated with the Cat.Net or Checkmarx static code analysis tools”.  Please add HP Fortify.
On Minimum Requirements, it says “required to run a TEAM Mentor.”  Delete the “a”

On Configuring Emails,  please change TM to TEAM Mentor.

On Using The Filters, it says “Pining Filters”.  It should be Pinning.

The 3 links at the bottom of Editing Overview are broken.

The links at the bottom of Editing WikiText Articles are broken.

The links at the bottom of Editing HTML  Articles are broken.

The links at the bottom of Editing Articles Meta-Data are broken.

The page TEAM Mentor Technology needs to be addressed.

The Eclipse Plugin header on the right should say “Eclipse for Fortify plugin” to be more descriptive.

On How to Evaluate and What to Expect (SCA section), you should put the link to the TM webpage https://www.securityinnovation.com/products/team-academy/secure-development-knowledgebase/

Change the link at the bottom of the Visual Studio FAQ page to https://www.securityinnovation.com/products/team-academy/secure-development-knowledgebase/ instead of Dinis’ blog.